First commit where the reading component actually works.
Added infrastructure to the particles for storing and retrieving data.
The data is stored according to the openPMD format, and converted to
the McStas format when a neutron is retrieved.

In openPMD_io two methods were added: init_read and trace_read, both
of which are included in the wrap.

The wrapped c compatible code also have a new way to initialize an
openPMD_io instance, as it needs to have READ_ONLY access instead of
CREATE.

In this version all openPMD data is loaded to the particles class
immediately, and this may not work with large datasets. It may be
necessary to load part of the data at a time using the chunk system.

Currently there seem to be a problem with the h5 file written by the
earlier code, as only the first chunk has data. It may be the first
chunk is overwritten many times, or only the first chunk is ever
written correctly.

More documentation and comments in the style as the existing code
will be added.